refactor: bounty info
