+++ site/src/components/Bounty/Detail/Meta/Info/InfoItem.js
